feat(footer): unify mono mega-footer, config-driven, site-wide
Pivot per user decision: upgrade the existing shared mono footer to the 9-column config-driven structure (footer-nav/footer-data) and render it once per marketing layout, replacing both the generic LandingFooter and the per-page MonoFooter. Fixes the double-footer on landing/models/providers/compare. New SiteMonoFooter is a self-contained .br-mono server component (semantic nav/h2). Retires components/landing/landing-footer.tsx.
