+++ components/models/mono-models-page.tsx
@@ -8,7 +8,6 @@
 import * as React from "react";
 import "../landing/mono/mono.css";
 import { Terminal, Ok, Dim, Faint } from "../landing/mono/terminal";
-import { MonoFooter } from "../landing/mono/landing";
 import { primeModelsCache, useModels, type Model } from "@/lib/models-data";
 import {
   CONTEXT_BUCKETS,
@@ -517,7 +516,6 @@ export function ModelsMonoPage({
   return (
     <div className="br-mono">
       <ModelsCatalog />
-      <MonoFooter />
     </div>
   );
 }
